Add Device trait and DeviceManager
Create Device trait to abstract basic device and bus behaviours.

Add DeviceManager to manage devices and resourses mapping.

This DeviceManager would contain all the buses and the mmio/pio operations
mapping with the corresponding address spaces.

Each real bus and device will make use of the method from DeviceManager
to allocate the space it needs. And also, it would register the mmio/pio
operations with the address space allocated before.

Signed-off-by: Jing Liu <jing2.liu@linux.intel.com>